add tags with depth integration tests and refactor test folder
