+++ tests/service/mod.rs
@@ -2,3 +2,5 @@ pub mod all;
 pub mod stream;
 pub mod tags;
 pub mod utils;
+pub mod user;
+pub mod endpoints;
