+++ tests/service/stream/post/kind/long.rs
@@ -1,31 +1,31 @@
 use super::DETROIT;
-use crate::service::stream::utils::verify_post_list_kind;
-use crate::service::stream::ROOT_PATH;
-use crate::service::stream::{AMSTERDAM, BOGOTA};
+use crate::service::stream::post::utils::verify_post_list_kind;
+use crate::service::stream::post::ROOT_PATH;
+use crate::service::stream::post::{AMSTERDAM, BOGOTA};
 use crate::service::utils::make_request;
 use anyhow::Result;
 
-const KIND: &str = "image";
+const KIND: &str = "long";
 
-const POST_I1: &str = "5YCW1TGL5BKG8";
-const POST_I2: &str = "5YCW1TGL5BKG7";
-const POST_I3: &str = "5YCW1TGL5BKG6";
-const POST_I4: &str = "5YCW1TGL5BKG5";
-const POST_I5: &str = "5YCW1TGL5BKG4";
-const POST_I6: &str = "5YCW1TGL5BKG3";
-const POST_I7: &str = "5YCW1TGL5BKG2";
-const POST_I8: &str = "5YCW1TGL5BKG1";
+const POST_L1: &str = "4ZCW1TGL5BKG8";
+const POST_L2: &str = "4ZCW1TGL5BKG7";
+const POST_L3: &str = "4ZCW1TGL5BKG6";
+const POST_L4: &str = "4ZCW1TGL5BKG5";
+const POST_L5: &str = "4ZCW1TGL5BKG4";
+const POST_L6: &str = "4ZCW1TGL5BKG3";
+const POST_L7: &str = "4ZCW1TGL5BKG2";
+const POST_L8: &str = "4ZCW1TGL5BKG1";
 
-pub const START_TIMELINE: &str = "1820477299345";
-pub const END_TIMELINE: &str = "1820477299308";
+pub const START_TIMELINE: &str = "1819477230345";
+pub const END_TIMELINE: &str = "1819477230308";
 
 #[tokio::test]
-async fn test_stream_image_post_kind() -> Result<()> {
-    let path = format!("{ROOT_PATH}?kind={KIND}");
+async fn test_stream_long_post_kind() -> Result<()> {
+    let path = format!("{ROOT_PATH}?kind=long");
 
     let body = make_request(&path).await?;
     let post_list = vec![
-        POST_I1, POST_I2, POST_I3, POST_I4, POST_I5, POST_I6, POST_I7, POST_I8,
+        POST_L1, POST_L2, POST_L3, POST_L4, POST_L5, POST_L6, POST_L7, POST_L8,
     ];
     verify_post_list_kind(post_list, body, KIND);
 
@@ -33,81 +33,81 @@ async fn test_stream_image_post_kind() -> Result<()> {
 }
 
 #[tokio::test]
-async fn test_stream_image_post_kind_with_start() -> Result<()> {
-    let path = format!("{ROOT_PATH}?kind={KIND}&start={START_TIMELINE}");
+async fn test_stream_long_post_kind_with_start() -> Result<()> {
+    let path = format!("{ROOT_PATH}?kind=long&start={START_TIMELINE}");
 
     let body = make_request(&path).await?;
-    let post_list = vec![POST_I3, POST_I4, POST_I5, POST_I6, POST_I7, POST_I8];
+    let post_list = vec![POST_L3, POST_L4, POST_L5, POST_L6, POST_L7, POST_L8];
     verify_post_list_kind(post_list, body, KIND);
 
     Ok(())
 }
 
 #[tokio::test]
-async fn test_stream_image_post_kind_with_end() -> Result<()> {
-    let path = format!("{ROOT_PATH}?kind={KIND}&end={START_TIMELINE}");
+async fn test_stream_long_post_kind_with_end() -> Result<()> {
+    let path = format!("{ROOT_PATH}?kind=long&end={START_TIMELINE}");
 
     let body = make_request(&path).await?;
-    let post_list = vec![POST_I1, POST_I2];
+    let post_list = vec![POST_L1, POST_L2];
     verify_post_list_kind(post_list, body, KIND);
 
     Ok(())
 }
 
 #[tokio::test]
-async fn test_stream_image_post_kind_with_start_and_end() -> Result<()> {
-    let path = format!("{ROOT_PATH}?kind={KIND}&start={START_TIMELINE}&end={END_TIMELINE}");
+async fn test_stream_long_post_kind_with_start_and_end() -> Result<()> {
+    let path = format!("{ROOT_PATH}?kind=long&start={START_TIMELINE}&end={END_TIMELINE}");
 
     let body = make_request(&path).await?;
-    let post_list = vec![POST_I3, POST_I4, POST_I5, POST_I6];
+    let post_list = vec![POST_L3, POST_L4, POST_L5, POST_L6];
     verify_post_list_kind(post_list, body, KIND);
 
     Ok(())
 }
 
 #[tokio::test]
-async fn test_stream_image_post_kind_with_author() -> Result<()> {
-    let path = format!("{ROOT_PATH}?kind={KIND}&author_id={BOGOTA}&source=author");
+async fn test_stream_long_post_kind_with_author() -> Result<()> {
+    let path = format!("{ROOT_PATH}?kind=long&author_id={BOGOTA}&source=author");
 
     let body = make_request(&path).await?;
-    let post_list = vec![POST_I3, POST_I5, POST_I8];
+    let post_list = vec![POST_L3, POST_L5, POST_L8];
     verify_post_list_kind(post_list, body, KIND);
 
     Ok(())
 }
 
 #[tokio::test]
-async fn test_stream_image_post_kind_with_author_skip_and_limit() -> Result<()> {
-    let path = format!("{ROOT_PATH}?kind={KIND}&author_id={BOGOTA}&source=author&skip=1&limit=1");
+async fn test_stream_long_post_kind_with_author_skip_and_limit() -> Result<()> {
+    let path = format!("{ROOT_PATH}?kind=long&author_id={BOGOTA}&source=author&skip=1&limit=1");
 
     let body = make_request(&path).await?;
-    let post_list = vec![POST_I5];
+    let post_list = vec![POST_L5];
     verify_post_list_kind(post_list, body, KIND);
 
     Ok(())
 }
 
-// ##### REACH: FOLLOWING ####
+// ##### REACH: FOLLOWERS ####
 
 #[tokio::test]
 async fn test_stream_post_kind_followers() -> Result<()> {
     let path = format!("{ROOT_PATH}?source=followers&observer_id={DETROIT}&kind={KIND}");
 
     let body = make_request(&path).await?;
-    let post_list = vec![POST_I1, POST_I2];
+    let post_list = vec![POST_L1, POST_L2];
     verify_post_list_kind(post_list, body, KIND);
 
     Ok(())
 }
 
-const REACH_START_TIMELINE: &str = "1820477299355";
+const REACH_START_TIMELINE: &str = "1819477230355";
 
 #[tokio::test]
 async fn test_stream_post_kind_followers_with_start() -> Result<()> {
     let path = format!("{ROOT_PATH}?source=followers&observer_id={DETROIT}&kind={KIND}&start={REACH_START_TIMELINE}");
 
     let body = make_request(&path).await?;
-    let post_list = vec![POST_I2];
+    let post_list = vec![POST_L2];
     verify_post_list_kind(post_list, body, KIND);
 
     Ok(())
@@ -120,7 +120,7 @@ async fn test_stream_post_kind_following() -> Result<()> {
     let path = format!("{ROOT_PATH}?source=following&observer_id={AMSTERDAM}&kind={KIND}");
 
     let body = make_request(&path).await?;
-    let post_list = vec![POST_I3, POST_I4, POST_I5, POST_I6, POST_I7, POST_I8];
+    let post_list = vec![POST_L3, POST_L4, POST_L5, POST_L6, POST_L7, POST_L8];
     verify_post_list_kind(post_list, body, KIND);
 
     Ok(())
@@ -129,11 +129,11 @@ async fn test_stream_post_kind_following() -> Result<()> {
 #[tokio::test]
 async fn test_stream_post_kind_following_with_start() -> Result<()> {
     let path = format!(
-        "{ROOT_PATH}?source=following&observer_id={AMSTERDAM}&kind={KIND}&start=1820477299325"
+        "{ROOT_PATH}?source=following&observer_id={AMSTERDAM}&kind={KIND}&start=1819477230325"
     );
 
     let body = make_request(&path).await?;
-    let post_list = vec![POST_I5, POST_I6, POST_I7, POST_I8];
+    let post_list = vec![POST_L5, POST_L6, POST_L7, POST_L8];
     verify_post_list_kind(post_list, body, KIND);
 
     Ok(())
@@ -146,7 +146,7 @@ async fn test_stream_post_kind_friends() -> Result<()> {
     let path = format!("{ROOT_PATH}?source=friends&observer_id={DETROIT}&kind={KIND}");
 
     let body = make_request(&path).await?;
-    let post_list = vec![POST_I1, POST_I2];
+    let post_list = vec![POST_L1, POST_L2];
     verify_post_list_kind(post_list, body, KIND);
 
     Ok(())
@@ -159,7 +159,7 @@ async fn test_stream_post_kind_friends_with_start() -> Result<()> {
     );
 
     let body = make_request(&path).await?;
-    let post_list = vec![POST_I2];
+    let post_list = vec![POST_L2];
     verify_post_list_kind(post_list, body, KIND);
 
     Ok(())
