+++ tests/service/stream/post/kind/mod.rs
