+++ tests/service/stream/post/kind/video.rs
@@ -1,31 +1,31 @@
 use super::DETROIT;
-use crate::service::stream::utils::verify_post_list_kind;
-use crate::service::stream::ROOT_PATH;
-use crate::service::stream::{AMSTERDAM, BOGOTA};
+use crate::service::stream::post::utils::verify_post_list_kind;
+use crate::service::stream::post::ROOT_PATH;
+use crate::service::stream::post::{AMSTERDAM, BOGOTA};
 use crate::service::utils::make_request;
 use anyhow::Result;
 
-const KIND: &str = "long";
+const KIND: &str = "video";
 
-const POST_L1: &str = "4ZCW1TGL5BKG8";
-const POST_L2: &str = "4ZCW1TGL5BKG7";
-const POST_L3: &str = "4ZCW1TGL5BKG6";
-const POST_L4: &str = "4ZCW1TGL5BKG5";
-const POST_L5: &str = "4ZCW1TGL5BKG4";
-const POST_L6: &str = "4ZCW1TGL5BKG3";
-const POST_L7: &str = "4ZCW1TGL5BKG2";
-const POST_L8: &str = "4ZCW1TGL5BKG1";
+const POST_V1: &str = "MLOW1TGL5BKH8";
+const POST_V2: &str = "MLOW1TGL5BKH7";
+const POST_V3: &str = "MLOW1TGL5BKH6";
+const POST_V4: &str = "MLOW1TGL5BKH5";
+const POST_V5: &str = "MLOW1TGL5BKH4";
+const POST_V6: &str = "MLOW1TGL5BKH3";
+const POST_V7: &str = "MLOW1TGL5BKH2";
+const POST_V8: &str = "MLOW1TGL5BKH1";
 
-pub const START_TIMELINE: &str = "1819477230345";
-pub const END_TIMELINE: &str = "1819477230308";
+pub const START_TIMELINE: &str = "1980477299345";
+pub const END_TIMELINE: &str = "1980477299308";
 
 #[tokio::test]
-async fn test_stream_long_post_kind() -> Result<()> {
-    let path = format!("{ROOT_PATH}?kind=long");
+async fn test_stream_video_post_kind() -> Result<()> {
+    let path = format!("{ROOT_PATH}?kind={KIND}");
 
     let body = make_request(&path).await?;
     let post_list = vec![
-        POST_L1, POST_L2, POST_L3, POST_L4, POST_L5, POST_L6, POST_L7, POST_L8,
+        POST_V1, POST_V2, POST_V3, POST_V4, POST_V5, POST_V6, POST_V7, POST_V8,
     ];
     verify_post_list_kind(post_list, body, KIND);
 
@@ -33,55 +33,55 @@ async fn test_stream_long_post_kind() -> Result<()> {
 }
 
 #[tokio::test]
-async fn test_stream_long_post_kind_with_start() -> Result<()> {
-    let path = format!("{ROOT_PATH}?kind=long&start={START_TIMELINE}");
+async fn test_stream_video_post_kind_with_start() -> Result<()> {
+    let path = format!("{ROOT_PATH}?kind={KIND}&start={START_TIMELINE}");
 
     let body = make_request(&path).await?;
-    let post_list = vec![POST_L3, POST_L4, POST_L5, POST_L6, POST_L7, POST_L8];
+    let post_list = vec![POST_V3, POST_V4, POST_V5, POST_V6, POST_V7, POST_V8];
     verify_post_list_kind(post_list, body, KIND);
 
     Ok(())
 }
 
 #[tokio::test]
-async fn test_stream_long_post_kind_with_end() -> Result<()> {
-    let path = format!("{ROOT_PATH}?kind=long&end={START_TIMELINE}");
+async fn test_stream_video_post_kind_with_end() -> Result<()> {
+    let path = format!("{ROOT_PATH}?kind={KIND}&end={START_TIMELINE}");
 
     let body = make_request(&path).await?;
-    let post_list = vec![POST_L1, POST_L2];
+    let post_list = vec![POST_V1, POST_V2];
     verify_post_list_kind(post_list, body, KIND);
 
     Ok(())
 }
 
 #[tokio::test]
-async fn test_stream_long_post_kind_with_start_and_end() -> Result<()> {
-    let path = format!("{ROOT_PATH}?kind=long&start={START_TIMELINE}&end={END_TIMELINE}");
+async fn test_stream_video_post_kind_with_start_and_end() -> Result<()> {
+    let path = format!("{ROOT_PATH}?kind={KIND}&start={START_TIMELINE}&end={END_TIMELINE}");
 
     let body = make_request(&path).await?;
-    let post_list = vec![POST_L3, POST_L4, POST_L5, POST_L6];
+    let post_list = vec![POST_V3, POST_V4, POST_V5, POST_V6];
     verify_post_list_kind(post_list, body, KIND);
 
     Ok(())
 }
 
 #[tokio::test]
-async fn test_stream_long_post_kind_with_author() -> Result<()> {
-    let path = format!("{ROOT_PATH}?kind=long&author_id={BOGOTA}&source=author");
+async fn test_stream_video_post_kind_with_author() -> Result<()> {
+    let path = format!("{ROOT_PATH}?kind={KIND}&author_id={BOGOTA}&source=author");
 
     let body = make_request(&path).await?;
-    let post_list = vec![POST_L3, POST_L5, POST_L8];
+    let post_list = vec![POST_V3, POST_V5, POST_V8];
     verify_post_list_kind(post_list, body, KIND);
 
     Ok(())
 }
 
 #[tokio::test]
-async fn test_stream_long_post_kind_with_author_skip_and_limit() -> Result<()> {
-    let path = format!("{ROOT_PATH}?kind=long&author_id={BOGOTA}&source=author&skip=1&limit=1");
+async fn test_stream_video_post_kind_with_author_skip_and_limit() -> Result<()> {
+    let path = format!("{ROOT_PATH}?kind={KIND}&author_id={BOGOTA}&source=author&skip=1&limit=1");
 
     let body = make_request(&path).await?;
-    let post_list = vec![POST_L5];
+    let post_list = vec![POST_V5];
     verify_post_list_kind(post_list, body, KIND);
 
     Ok(())
@@ -94,20 +94,20 @@ async fn test_stream_post_kind_followers() -> Result<()> {
     let path = format!("{ROOT_PATH}?source=followers&observer_id={DETROIT}&kind={KIND}");
 
     let body = make_request(&path).await?;
-    let post_list = vec![POST_L1, POST_L2];
+    let post_list = vec![POST_V1, POST_V2];
     verify_post_list_kind(post_list, body, KIND);
 
     Ok(())
 }
 
-const REACH_START_TIMELINE: &str = "1819477230355";
+const REACH_START_TIMELINE: &str = "1980477299355";
 
 #[tokio::test]
 async fn test_stream_post_kind_followers_with_start() -> Result<()> {
     let path = format!("{ROOT_PATH}?source=followers&observer_id={DETROIT}&kind={KIND}&start={REACH_START_TIMELINE}");
 
     let body = make_request(&path).await?;
-    let post_list = vec![POST_L2];
+    let post_list = vec![POST_V2];
     verify_post_list_kind(post_list, body, KIND);
 
     Ok(())
@@ -120,7 +120,7 @@ async fn test_stream_post_kind_following() -> Result<()> {
     let path = format!("{ROOT_PATH}?source=following&observer_id={AMSTERDAM}&kind={KIND}");
 
     let body = make_request(&path).await?;
-    let post_list = vec![POST_L3, POST_L4, POST_L5, POST_L6, POST_L7, POST_L8];
+    let post_list = vec![POST_V3, POST_V4, POST_V5, POST_V6, POST_V7, POST_V8];
     verify_post_list_kind(post_list, body, KIND);
 
     Ok(())
@@ -129,11 +129,11 @@ async fn test_stream_post_kind_following() -> Result<()> {
 #[tokio::test]
 async fn test_stream_post_kind_following_with_start() -> Result<()> {
     let path = format!(
-        "{ROOT_PATH}?source=following&observer_id={AMSTERDAM}&kind={KIND}&start=1819477230325"
+        "{ROOT_PATH}?source=following&observer_id={AMSTERDAM}&kind={KIND}&start=1980477299325"
     );
 
     let body = make_request(&path).await?;
-    let post_list = vec![POST_L5, POST_L6, POST_L7, POST_L8];
+    let post_list = vec![POST_V5, POST_V6, POST_V7, POST_V8];
     verify_post_list_kind(post_list, body, KIND);
 
     Ok(())
@@ -146,7 +146,7 @@ async fn test_stream_post_kind_friends() -> Result<()> {
     let path = format!("{ROOT_PATH}?source=friends&observer_id={DETROIT}&kind={KIND}");
 
     let body = make_request(&path).await?;
-    let post_list = vec![POST_L1, POST_L2];
+    let post_list = vec![POST_V1, POST_V2];
     verify_post_list_kind(post_list, body, KIND);
 
     Ok(())
@@ -159,7 +159,7 @@ async fn test_stream_post_kind_friends_with_start() -> Result<()> {
     );
 
     let body = make_request(&path).await?;
-    let post_list = vec![POST_L2];
+    let post_list = vec![POST_V2];
     verify_post_list_kind(post_list, body, KIND);
 
     Ok(())
