+++ tests/service/stream/post/reach/engagement.rs
@@ -1,6 +1,6 @@
 use super::utils::test_reach_filter_with_posts;
-use crate::service::stream::{AMSTERDAM, BOGOTA, ROOT_PATH};
-use crate::service::{stream::TAG_LABEL_2, utils::make_wrong_request};
+use crate::service::stream::post::{AMSTERDAM, BOGOTA, ROOT_PATH};
+use crate::service::{stream::post::TAG_LABEL_2, utils::make_wrong_request};
 use anyhow::Result;
 
 // ›››››› THE BELLOW REQUESTS HITS THE GRAPH ‹‹‹‹‹‹‹
