+++ tests/service/stream/user/mod.rs
@@ -1,0 +1,4 @@
+pub mod list;
+pub mod score;
+pub mod search;
+pub mod reach;
