+++ tests/service/user/mod.rs
@@ -1,0 +1,3 @@
+pub mod reach;
+pub mod views;
+pub mod search;
